rune/libenclave/skeleton: Rebase pal api version 1
